backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-core==2.14.1
uvicorn==0.24.0
pyjwt==2.8.0
python-jose==3.3.0
cryptography==41.0.8
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
better-auth==0.0.1-beta.9